requirements.txt
pyyaml >= 5.1
visualdl >= 2.2.0
opencv-python
tqdm
filelock
scipy
prettytable
sklearn == 0.0

Flask
FLask-APScheduler
Flask-Migrate
Flask-Login
Flask-Reuploaded
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
PyMySQL
psutil
Flask-Mail
Pillow
validators
cryptography
python-dotenv
sqlparse
